Add offset-aware ords block loaders for multi_value support
Wires an offsets doc value into the ords-based BytesRef block loaders
(BytesRefsFromOrdsBlockLoader and its Mv variants) so multi-value docs
read in indexing order. Adds BlockFactory.arrayOrderOrdinalsBuilder and
backs it with ArrayOrderOrdinalsBuilder, factored out alongside
SortedSetOrdinalsBuilder via a shared OrdinalsBuilderBase.
