+++ server/src/test/java/org/elasticsearch/index/mapper/blockloader/docvalues/AbstractArrayOrderBlockLoaderTests.java
@@ -126,7 +126,7 @@ public abstract class AbstractArrayOrderBlockLoaderTests<V extends Comparable<V>
                     // alternate shape (Singleton, BytesRefsFromBinary, ...) it lands on instead
                     assertFalse(
                         "dispatcher must elide ArrayOrder for single-valued segments, got " + r,
-                        r instanceof AbstractNumericBlockLoader.ArrayOrder<?>
+                        r instanceof AbstractNumericBlockLoader.ArrayOrder<?> || r instanceof BytesRefsFromOrdsBlockLoader.ArrayOrder
                     );
                     TestBlock block = (TestBlock) r.read(TestBlock.factory(), TestBlock.docs(0, 1), 0, false);
                     assertEquals(val1, block.get(0));
