+++ test/framework/src/main/java/org/elasticsearch/index/mapper/TestBlock.java
@@ -458,6 +458,15 @@ public class TestBlock implements BlockLoader.Block {
 
             @Override
             public BlockLoader.SortedSetOrdinalsBuilder sortedSetOrdinalsBuilder(SortedSetDocValues ordinals, int expectedSize) {
+                return newSortedSetOrdinalsBuilder(ordinals, expectedSize);
+            }
+
+            @Override
+            public BlockLoader.SortedSetOrdinalsBuilder arrayOrderOrdinalsBuilder(SortedSetDocValues ordinals, int expectedSize) {
+                return newSortedSetOrdinalsBuilder(ordinals, expectedSize);
+            }
+
+            private BlockLoader.SortedSetOrdinalsBuilder newSortedSetOrdinalsBuilder(SortedSetDocValues ordinals, int expectedSize) {
                 class SortedSetOrdinalBuilder extends TestBlock.Builder implements BlockLoader.SortedSetOrdinalsBuilder {
                     private SortedSetOrdinalBuilder() {
                         super(expectedSize);
